requirement.txt
colorama==0.4.6
paho-mqtt==2.1.0
python-dotenv==1.1.0
